// src/files/dto/upload-instructions.dto.ts
import { FileVisibility } from '../../../generated/prisma/enums';
import { ApiProperty } from '@nestjs/swagger';
import { ApiPropertyOptional } from '@nestjs/swagger';

export class UploadInstructionsDto {
  @ApiProperty({
    description: 'Internal file record id.',
    example: 'd8d89d76-1a4c-4cc8-b804-e7fcf58567af',
  })
  fileId!: string;

  @ApiProperty({
    description: 'Object key generated by backend.',
    example:
      'users/user-1/avatar/2026-04-05/6f302327-9738-44cb-b469-5614f0c6f966.png',
  })
  key!: string;

  @ApiProperty({
    enum: FileVisibility,
    example: FileVisibility.PRIVATE,
    description: 'Visibility selected for the file.',
  })
  visibility!: FileVisibility;

  @ApiProperty({
    description:
      'Temporary presigned URL for direct browser-to-storage upload. Send an HTTP PUT request to this URL with the raw file bytes as the request body.',
    example:
      'https://93a195452e08956a31bcbd3b8d185d5f.r2.cloudflarestorage.com/nti-bucket/users/fae01d73-b528-4af1-9765-f22ba9cf6e02/general/2026-04-08/dc1f54c3-8bd3-41a1-ae9f-5eb039715c1f.png?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Expires=120&x-id=PutObject',
  })
  uploadUrl!: string;

  @ApiPropertyOptional({
    description:
      'Stable public URL for reading the file when visibility is PUBLIC. This field is omitted for PRIVATE files.',
    example:
      'https://cdn.example.com/users/fae01d73-b528-4af1-9765-f22ba9cf6e02/avatar/2026-04-08/dc1f54c3-8bd3-41a1-ae9f-5eb039715c1f.png',
  })
  publicUrl?: string;

  @ApiProperty({
    description: 'ISO timestamp when upload URL expires.',
    example: '2026-04-05T11:51:10.000Z',
  })
  expiresAt!: string;
}